Uploaded version W0.0.2P
Version W0.0.2P checks the time and day of the week on run, and opens the correct class link if appropriate. Schedule is fixed to WHS 2020-2021 schedule and links must be configured by editing the file. Compatible with firefox only. All of these things will be remedied in planned future versions.